feat(browser): Set page context instead of using request interface
