+++ dev-packages/browser-integration-tests/suites/integrations/httpContext/test.ts
@@ -4,7 +4,7 @@ import type { Event } from '@sentry/core';
 import { sentryTest } from '../../../utils/fixtures';
 import { getFirstSentryEnvelopeRequest } from '../../../utils/helpers';
 
-sentryTest('httpContextIntegration captures user-agent and referrer', async ({ getLocalTestUrl, page }) => {
+sentryTest('pageInformationIntegration captures user-agent and referrer', async ({ getLocalTestUrl, page }) => {
   const url = await getLocalTestUrl({ testDir: __dirname });
 
   const errorEventPromise = getFirstSentryEnvelopeRequest<Event>(page);
